Project6_Telegram_chatbot_using_openAI/requirements.txt
python-telegram-bot
openai==1.63.2
python-dotenv